gpu memory does not release
