+++ tests/flux/test_flux_examples.py
@@ -1,7 +1,9 @@
+import gc
 import os
 import subprocess
 
 import pytest
+import torch
 
 EXAMPLES_DIR = "./examples"
 
@@ -10,6 +12,8 @@ example_scripts = [f for f in os.listdir(EXAMPLES_DIR) if f.endswith(".py") and 
 
 @pytest.mark.parametrize("script_name", example_scripts)
 def test_example_script_runs(script_name):
+    gc.collect()
+    torch.cuda.empty_cache()
     script_path = os.path.join(EXAMPLES_DIR, script_name)
     result = subprocess.run(["python", script_path], capture_output=True, text=True)
     print(f"Running {script_path} -> Return code: {result.returncode}")
